fixed build

+++ Software/Arduino code/OpenAstroTracker/src/inc/Globals.cpp
@@ -4,4 +4,4 @@
 // TODO (AS): This is just a tmp fix, we need to get rid of these globals
 
 float RAStepsPerDegree = RA_STEPS_PER_DEGREE;
-int DECStepsPerDegree = DEC_STEPS_PER_DEGREE;
+float DECStepsPerDegree = DEC_STEPS_PER_DEGREE;
